Update README.md
Removed any info pertaining to CONTRIBUTING. Those will be placed in a separate CONTRIBUTING.md. Removed the link to the Prototype Kickoff, since it was outdated. Replaced the link to the Prototype Kickoff with a link to the "Welcome! Read me first!" document on Google Drive. Added a link to the UI/UX mockup. Added a section for Maintainers so potential contributers know who to message first.